refactor(utils): move {exec, execAsync,spawn}()
