+++ scripts/release/utils.ts
@@ -1,50 +1,7 @@
 /* This file is a part of @mdn/browser-compat-data
  * See LICENSE file for more information. */
 
-import {
-  execSync,
-  ExecSyncOptionsWithStringEncoding,
-  spawnSync,
-  SpawnSyncOptionsWithStringEncoding,
-} from 'node:child_process';
-
-/**
- * Execute a command
- * @param command The command to execute
- * @param opts The options to pass to execSync
- * @returns The output from the command
- */
-export const exec = (
-  command: string,
-  opts?: ExecSyncOptionsWithStringEncoding,
-): string => execSync(command, { encoding: 'utf8', ...opts }).trim();
-
-/**
- * Execute a command
- * @param command The command to execute
- * @param args The arguments to pass
- * @param opts The options to pass to spawnSync
- * @returns The output from the command
- */
-export const spawn = (
-  command: string,
-  args: readonly string[],
-  opts?: SpawnSyncOptionsWithStringEncoding,
-): string => {
-  const result = spawnSync(command, args, { encoding: 'utf8', ...opts });
-
-  if (result.error) {
-    throw result.error;
-  }
-
-  if (result.status !== 0) {
-    throw new Error(
-      `The command '${command}' returned non-zero exit status ${result.status}: ${result.stderr}`,
-    );
-  }
-
-  return result.stdout.trim();
-};
+import { exec } from '../../utils/index.js';
 
 /**
  * Check for GitHub CLI and exit the program if it's not existent
@@ -52,7 +9,7 @@ export const spawn = (
 export const requireGitHubCLI = (): void => {
   const command = 'gh auth status';
   try {
-    execSync(command, { encoding: 'utf8', stdio: 'ignore' });
+    exec(command, { stdio: 'ignore' });
   } catch (err) {
     console.trace(err);
     console.error(`Error: ${command} failed.`);
