+++ utils/index.ts
@@ -1,9 +1,12 @@
 /* This file is a part of @mdn/browser-compat-data
  * See LICENSE file for more information. */
 
+import exec from './exec.js';
+import execAsync from './exec-async.js';
 import iterSupport from './iter-support.js';
 import query from './query.js';
+import spawn from './spawn.js';
 import walk from './walk.js';
 import normalizePath from './normalize-path.js';
 
-export { iterSupport, normalizePath, query, walk };
+export { exec, execAsync, iterSupport, normalizePath, query, spawn, walk };
